+++ src/Reveries.Api/Configuration/HealthCheck/HealthCheckExtensions.cs
@@ -10,7 +10,7 @@ public static class HealthCheckExtensions
     public static IServiceCollection AddApplicationHealthChecks(this IServiceCollection services)
     {
         services.AddHealthChecks()
-            .AddPostgresqlHealthCheck()
+            .AddPostgresHealthCheck()
             .AddRedisHealthCheck();
         
         return services;
