+++ src/Reveries.Infrastructure/DependencyInjection.cs
@@ -9,7 +9,7 @@ public static class DependencyInjection
 {
     public static IServiceCollection AddInfrastructure(this IServiceCollection services, IConfiguration config)
     {
-        services.AddPostgresql(config);
+        services.AddPostgres(config);
         services.AddRedisCache(config);
         
         return services;
