+++ src/Reveries.Persistence/Repositories/DeweyDecimalsRepository.cs
@@ -1,6 +1,7 @@
 using Dapper;
 using Reveries.Domain.Interfaces.IRepository;
 using Reveries.Domain.ValueObjects;
+using Reveries.Persistence.Context;
 using Reveries.Persistence.Interfaces;
 using Reveries.Persistence.Mappers;
